feat(no-milestone): add an option to not use milestones
For my own workflow, it is also useful to move tasks around on
our main board and not just on milestones. The --no-milestone
option allows to do just that.

Some of the logic to select project or milestone and to find a
column on a workboard has moved to phabricator.py as it seems
more readable.