Introduce in-memory/persistent caching
Some resources, such as projects or users won't change much over time,
meaning that they should be safe to cache on disk, to speed up subsequent
invocations.

We load the cache at startup, and dump the in-memory cache to disk at exit.

This has shown to divide the runtime of a task show by ~2 on the second
run.