feat(move): Project-PHID as a CLI argument for the move command
When working with multiple projects, it is cumbersome to edit
config file or environment variables each time a different
project is needed. This commit introduces a "--project-phid" argument
to the "move" command, which would override
"phabricator_default_project_phid".

It might make sense in a next step to introduce the same
"--project-phid" argument to other commands where it makes sense.