fix(PhabricatorClient.find_tasks): don't crash on no columns
As described in the linked issue #10, when find_tasks is passed
`columns=None` it crashes as it can't iterate `None`.

`find_tasks()` now marks the `columns_phids` argument as a
non optional sequence. This documents that this argument is
expected to not be modified and that `None` should not be passed.

Caller is modified to follow this new behaviour.

fixes: #10